pyproject.toml
[project]
name = "vsc-release-note-summary"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "prompt-toolkit>=3.0.52",
    "strands-agents[ollama]>=1.19.0",
]
